+++ src/Instance/InstanceDetails/InstanceDescriptiveView/utils.js
@@ -1,8 +1,8 @@
-import languagetable from '../../../data/languages';
+import { formattedLanguageName } from '@folio/stripes/components';
 
 // eslint-disable-next-line
-export const formatLanguages = (languages = []) => {
+export const formatLanguages = (languages = [], intl) => {
   return languages
-    .map(languageCode => languagetable.languageByCode(languageCode))
+    .map(languageCode => formattedLanguageName(languageCode, intl))
     .join(', ');
 };