+++ src/Instance/InstanceMovement/InstanceMovementDetails/InstanceMovementDetails.js
@@ -5,6 +5,10 @@ import React, {
 import PropTypes from 'prop-types';
 
 import { Droppable } from 'react-beautiful-dnd';
+import {
+  useStripes,
+} from '@folio/stripes/core';
+
 import DataContext from '../../../contexts/DataContext';
 import {
   InstanceDetails,
@@ -16,6 +20,8 @@ import {
 import InstanceMovementDetailsActions from './InstanceMovementDetailsActions';
 
 const InstanceMovementDetails = ({ instance, onClose, hasMarc }) => {
+  const stripes = useStripes();
+
   const closeInstance = useCallback(() => {
     onClose(instance);
   }, [instance, onClose]);
@@ -26,7 +32,10 @@ const InstanceMovementDetails = ({ instance, onClose, hasMarc }) => {
   } = useContext(DataContext);
 
   const getActionMenu = useCallback(({ onToggle }) => {
-    if (instance.source !== 'MARC') {
+    if (
+      instance.source !== 'MARC'
+      && !stripes.hasPerm('inventory.instances.item.put')
+    ) {
       return null;
     }