+++ src/ViewInstance.js
@@ -9,27 +9,17 @@ import PropTypes from 'prop-types';
 import ReactRouterPropTypes from 'react-router-prop-types';
 import { FormattedMessage } from 'react-intl';
 
-import queryString from 'query-string';
-
 import {
   AppIcon,
-  IntlConsumer,
   IfPermission,
 } from '@folio/stripes/core';
 import {
   Pane,
   Button,
-  Layer,
   Icon,
   Callout,
 } from '@folio/stripes/components';
 
-import {
-  craftLayerUrl,
-  marshalInstance,
-  unmarshalInstance,
-} from './utils';
-import InstanceForm from './edit/InstanceForm';
 import ViewHoldingsRecord from './ViewHoldingsRecord';
 import makeConnectedInstance from './ConnectedInstance';
 import withLocation from './withLocation';
@@ -98,21 +88,37 @@ class ViewInstance extends React.Component {
     this.instanceId = null;
     this.cViewHoldingsRecord = this.props.stripes.connect(ViewHoldingsRecord);
 
-    this.craftLayerUrl = craftLayerUrl.bind(this);
     this.calloutRef = createRef();
   }
 
   componentDidMount() {
-    this.getMARCRecord();
+    const isMARCSource = this.isMARCSource();
+
+    if (isMARCSource) {
+      this.getMARCRecord();
+    }
   }
 
   componentDidUpdate(prevProps) {
-    const { location: prevLocation } = prevProps;
-    const { location } = this.props;
+    const { resources: prevResources } = prevProps;
+    const { resources } = this.props;
+    const instanceRecords = resources?.selectedInstance?.records;
+    const instanceRecordsId = instanceRecords[0]?.id;
+    const prevInstanceRecordsId = prevResources?.selectedInstance?.records[0]?.id;
+    const isMARCSource = this.isMARCSource();
 
-    if (prevLocation === location) return;
-    this.getMARCRecord();
+    if (isMARCSource && instanceRecordsId !== prevInstanceRecordsId) {
+      this.getMARCRecord();
+    }
   }
+
+  isMARCSource = () => {
+    const { resources } = this.props;
+    const instanceRecords = resources?.selectedInstance?.records;
+    const instanceRecordsSource = instanceRecords?.[0]?.source;
+
+    return instanceRecordsSource === 'MARC';
+  };
 
   getMARCRecord = () => {
     const { mutator } = this.props;
@@ -125,9 +131,14 @@ class ViewInstance extends React.Component {
   };
 
   // Edit Instance Handlers
-  onClickEditInstance = (e) => {
-    if (e) e.preventDefault();
-    this.props.updateLocation({ layer: 'edit' });
+  onClickEditInstance = () => {
+    const { history, location, match } = this.props;
+    const instanceId = match.params.id;
+
+    history.push({
+      pathname: `/inventory/edit/${instanceId}/instance`,
+      search: location.search,
+    });
   };
 
   editInstanceMarc = () => {
@@ -181,22 +192,6 @@ class ViewInstance extends React.Component {
 
         this.calloutRef.current.sendCallout({ message });
       });
-  };
-
-  update = (instance) => {
-    const { referenceTables: { identifierTypesByName } } = this.props;
-
-    // Massage record to add preceeding and succeeding title fields
-    marshalInstance(instance, identifierTypesByName);
-
-    this.props.mutator.selectedInstance.PUT(instance).then(() => {
-      this.resetLayerQueryParam();
-    });
-  };
-
-  resetLayerQueryParam = (e) => {
-    if (e) e.preventDefault();
-    this.props.updateLocation({ layer: null });
   };
 
   goBack = (e) => {
@@ -263,7 +258,6 @@ class ViewInstance extends React.Component {
         <IfPermission perm="ui-inventory.instance.edit">
           <Button
             id="edit-instance"
-            href={this.craftLayerUrl('edit')}
             onClick={() => {
               onToggle();
               this.onClickEditInstance();
@@ -362,16 +356,12 @@ class ViewInstance extends React.Component {
   render() {
     const {
       match: { params: { id, holdingsrecordid, itemid } },
-      location,
       referenceTables,
       stripes,
       onClose,
       paneWidth,
     } = this.props;
 
-    const { identifierTypesById } = referenceTables;
-
-    const query = location.search ? queryString.parse(location.search) : {};
     const ci = makeConnectedInstance(this.props, stripes.logger);
     const instance = ci.instance();
 
@@ -392,29 +382,6 @@ class ViewInstance extends React.Component {
             />
           </div>
         </Pane>
-      );
-    }
-
-    if (query.layer === 'edit') {
-      return (
-        <IntlConsumer>
-          {(intl) => (
-            <Layer
-              isOpen
-              contentLabel={intl.formatMessage({ id: 'ui-inventory.editInstanceDialog' })}
-            >
-              <InstanceForm
-                onSubmit={this.update}
-                initialValues={unmarshalInstance(instance, identifierTypesById)}
-                instanceSource={get(instance, ['source'])}
-                referenceTables={referenceTables}
-                stripes={stripes}
-                match={this.props.match}
-                onCancel={this.resetLayerQueryParam}
-              />
-            </Layer>
-          )}
-        </IntlConsumer>
       );
     }