+++ src/routes/index.js
@@ -5,3 +5,4 @@ export { default as QuickMarcRoute } from './QuickMarcRoute';
 export { default as CreateItemRoute } from './CreateItemRoute';
 export { default as CreateHoldingRoute } from './CreateHoldingRoute';
 export { default as InstanceMarcRoute } from './InstanceMarcRoute';
+export { default as InstanceEditRoute } from './InstanceEditRoute';